Initial attempt to support kennels that have legacy run counts without attendance data.  This was added for NOH3.
If an entry exists in the LEGACY_HASHINGS_TABLE for a hasher the number
of legacy runs will be added to the total run counts for that hasher.

This makes some things wonky, like "quickest hashers to 5
runs"...obviously the legacy data is ignore for stats like this that
require real attendance data.