Repair CSRF functionality.
This enables CSRF handling, sort of.  It works automagically with anything
that uses the form.factory - which is only three places:

1) password change form
2) new hasher form
3) modify hasher form

Added utilities in BaseController to allow CSRF tokens to be placed
other places when the form.factory isn't being used.  Currently this is
only one place - because it was already half wired up anyway - the delete
hasher page.

But now that the utils are here it should be easy to wire this up
everywhere.