Update loadUserByUsername query.
Only select values that are used (not *).

Remove strtolower on bind variable.  This removes a potential bug which
could occur if usernames are mixed case in the database and the collation
of the users table is case sensitive.  Since we don't set the collation
of the users table when it is created we rely on the database defaults,
which could be either.