Update to pure symfony 6.2.
Sort of.

Only the home page works, and right now it it sitting in the /public
directory.  Not sure what this means long term but for development
it is fine.

When this is done all the code in github will be application
code - no more dealing with framework related issues and trying to
maintain the old silex capabilities.

Took me awhile to see that this was the right way forward.